requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Data processing
pandas>=2.1.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
duckdb>=0.9.0
pyarrow>=14.0.0

# String matching
rapidfuzz>=3.5.0

# HTTP client
httpx>=0.26.0

# Environment variables
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Authentication
passlib[bcrypt]>=1.7.4
# Replacing incorrect 'jose' (Python2-era) package with maintained python-jose
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
